+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2022/07/20
+### Added
+- Add `DdsConversionOptions` to make mipmaps and shuffling configurable.
+### Changed
+- Minimum image width and height are now 8 pixels.
+### Fixed
+- Add `Bitmap` interface to documentation index.
+
 ## [0.2.0] - 2022/06/21
 ### Changed
 - Add options to DdsImage's `toShuffled()` and `toUnshuffled()`.